requirements.txt
pennylane
pennylane-qiskit

torch
torchvision
torchaudio
torcheval

qiskit
qiskit-iqm
qiskit-ibm-runtime
qiskit-machine-learning
qiskit_qasm3_import

scikit-image

niapy
mlflow
tqdm
pylatexenc